More refactoring to adapter test classes. Upgrade model adding guide.
