+++ tests_adapters/test_mbart.py
@@ -1,7 +1,15 @@
+import unittest
+
 from tests.test_modeling_mbart import *
 from transformers import MBartAdapterModel
 from transformers.testing_utils import require_torch
 
+from .test_adapter import AdapterTestBase, make_config
+from .test_adapter_common import AdapterModelTestMixin
+from .test_adapter_composition import ParallelAdapterInferenceTestMixin
+from .test_adapter_conversion import ModelClassConversionTestMixin
+from .test_adapter_fusion_common import AdapterFusionModelTestMixin
+from .test_adapter_heads import PredictionHeadModelTestMixin
 from .test_common import AdapterModelTesterMixin
 
 
@@ -10,3 +18,38 @@ class MBartAdapterModelTest(AdapterModelTesterMixin, MBartModelTest):
     all_model_classes = (
         MBartAdapterModel,
     )
+
+
+class MBartAdapterTestBase(AdapterTestBase):
+    config_class = MBartConfig
+    config = make_config(
+        MBartConfig,
+        d_model=16,
+        encoder_layers=2,
+        decoder_layers=2,
+        encoder_attention_heads=4,
+        decoder_attention_heads=4,
+        encoder_ffn_dim=4,
+        decoder_ffn_dim=4,
+    )
+
+
+@require_torch
+class MBartAdapterTest(
+    AdapterModelTestMixin,
+    AdapterFusionModelTestMixin,
+    PredictionHeadModelTestMixin,
+    ParallelAdapterInferenceTestMixin,
+    MBartAdapterTestBase,
+    unittest.TestCase,
+):
+    pass
+
+
+@require_torch
+class MBartClassConversionTest(
+    ModelClassConversionTestMixin,
+    MBartAdapterTestBase,
+    unittest.TestCase,
+):
+    pass
