+++ tests_adapters/test_roberta.py
@@ -1,7 +1,15 @@
+import unittest
+
 from tests.test_modeling_roberta import *
 from transformers import RobertaAdapterModel
 from transformers.testing_utils import require_torch
 
+from .test_adapter import AdapterTestBase, make_config
+from .test_adapter_common import AdapterModelTestMixin
+from .test_adapter_composition import ParallelAdapterInferenceTestMixin
+from .test_adapter_conversion import ModelClassConversionTestMixin
+from .test_adapter_fusion_common import AdapterFusionModelTestMixin
+from .test_adapter_heads import PredictionHeadModelTestMixin
 from .test_common import AdapterModelTesterMixin
 
 
@@ -10,3 +18,35 @@ class RobertaAdapterModelTest(AdapterModelTesterMixin, RobertaModelTest):
     all_model_classes = (
         RobertaAdapterModel,
     )
+
+
+class RobertaAdapterTestBase(AdapterTestBase):
+    config_class = RobertaConfig
+    config = make_config(
+        RobertaConfig,
+        hidden_size=32,
+        num_hidden_layers=4,
+        num_attention_heads=4,
+        intermediate_size=37,
+    )
+
+
+@require_torch
+class RobertaAdapterTest(
+    AdapterModelTestMixin,
+    AdapterFusionModelTestMixin,
+    PredictionHeadModelTestMixin,
+    ParallelAdapterInferenceTestMixin,
+    RobertaAdapterTestBase,
+    unittest.TestCase,
+):
+    pass
+
+
+@require_torch
+class RobertaClassConversionTest(
+    ModelClassConversionTestMixin,
+    RobertaAdapterTestBase,
+    unittest.TestCase,
+):
+    pass
